Changed readme in installation directory
The DISORT readme was renamed readme_disort.txt and the readme created
for this directory was renamed readme.txt. This was done so that the
readme created for this directory will show up when the github folder
“installation” is opened, rather than the DISORT readme, which is
specific only to the Stamnes et al. DISORT files, and does not describe
other files in this directory.